+++ backend/canisters/local_user_index/api/src/updates/c2c_delete_community.rs
@@ -1,16 +1,10 @@
 use candid::CandidType;
-use oc_error_codes::OCError;
 use serde::{Deserialize, Serialize};
-use types::CommunityId;
+use types::{CommunityId, UnitResult};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub community_id: CommunityId,
 }
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub enum Response {
-    Success,
-    CommunityNotFound,
-    Error(OCError),
-}
+pub type Response = UnitResult;