+++ backend/canisters/local_user_index/api/src/updates/c2c_group_canister.rs
@@ -1,13 +1,10 @@
 use crate::GroupEvent;
 use serde::{Deserialize, Serialize};
-use types::IdempotentEnvelope;
+use types::{IdempotentEnvelope, SuccessOnly};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Args {
     pub events: Vec<IdempotentEnvelope<GroupEvent>>,
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum Response {
-    Success,
-}
+pub type Response = SuccessOnly;